Added state for cards, renamed files
