+++ src/App.js
@@ -1,7 +1,7 @@
 import './App.css';
-import Main from './components/main/main';
+import Main from './components/main/Main';
 import '@fontsource/roboto/500.css';
-import Header from './components/header/header';
+import Header from './components/header/Header';
 
 function App() {
   return (
